requirements.txt
numpy==1.17.3
Pillow==8.2.0
torch==1.4.0+cpu
torchvision==0.5.0+cpu
scikit-learn==0.20.3
pandas==0.23.4